# documentation/swagger.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Covid19 Greek Data API
servers:
- url: http://95.216.160.230/api/
- url: http://localhost:8080
paths:
  /health:
    get:
      summary: simple healthcheck
      tags:
      - helpers
      responses:
        '200':
          description: OK
  /timeline_fields:
    get:
      summary: get all filter fields for /timeline endpoint
      tags:
      - helpers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/timelineFields'
  /regional_units:
    get:
      summary: Greece's prefecture geographical information
      tags:
      - geographical
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/regionalUnits'
  /municipalities:
    get:
      summary: Greece's municipality geographical information
      tags:
      - geographical
      parameters:
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/per_page'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/municipalities'
  /cases:
    get:
      summary: covid19 cases per prefecture
      tags:
      - covid19
      parameters:
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/per_page'
      - in: query
        name: regional_unit_id
        schema:
          description: the id of a specific geographical region (see /regional_units endpoint)
          type: integer
          example: 2
      - in: query
        name: start_date
        schema:
          description: the first date of the cases period
          type: string
          example: 2022-01-01
      - in: query
        name: end_date
        schema:
          description: the last date of the cases period
          type: string
          example: 2022-01-10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/municipalityCasesList'
  /deaths_per_municipality:
    get:
      summary: covid19 deaths per municipality
      tags:
      - covid19
      parameters:
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/per_page'
      - in: query
        name: municipality_id
        schema:
          description: the id of a specific municipality (see /municipalities endpoint)
          type: integer
          example: 12
      - in: query
        name: year
        schema:
          description: a specific year of deaths
          type: integer
          example: 2021
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/deathsPerMunicipality'
  /timeline:
    get:
      summary: get full covid19 info for every date at a specific period
      tags:
      - covid19
      parameters:
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/per_page'
      - in: query
        name: fields
        schema:
          description: filter specific fields, separated by comma (see GET /timeline_fields endpoint for a full list)
          type: string
          example: intubated,deaths
      - in: query
        name: start_date
        schema:
          description: the first date of the timeline period
          type: string
          example: 2022-01-01
      - in: query
        name: end_date
        schema:
          description: the last date of the timeline period
          type: string
          example: 2022-01-10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/covidInfo'
  /{field}:
    get:
      summary: get covid19 timeline info for a specific field (same as GET /timeline?fields={field})
      tags:
      - covid19
      parameters:
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/per_page'
      - in: path
        name: field
        required: true
        schema:
          type: string
          example: intubated
      - in: query
        name: start_date
        schema:
          description: the first date of the timeline period
          type: string
          example: 2022-01-01
      - in: query
        name: end_date
        schema:
          description: the last date of the timeline period
          type: string
          example: 2022-01-10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/singleFieldCovidInfo'
  /demographics:
    get:
      summary: get covid19 demographic info per date and per age category
      tags:
      - covid19
      parameters:
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/per_page'
      - in: query
        name: category
        schema:
          description: filter by specific category. Available categories are 0-17,18-39-40-64,65+
          type: string
          example: 18-39
      - in: query
        name: start_date
        schema:
          description: the first date of the timeline period
          type: string
          example: 2022-01-01
      - in: query
        name: end_date
        schema:
          description: the last date of the timeline period
          type: string
          example: 2022-01-10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/demographicInfo'
components:
  schemas:
    municipalityCases:
      description: cases per municipality at a specific date
      type: object
      properties:
        regional_unit_id:
          type: integer
          example: 1
        date:
          type: string
          example: "2022-01-01"
        cases:
          type: integer
          example: 143
    municipalityCasesList:
      type: array
      items:
        $ref: '#/components/schemas/municipalityCases'
    timelineFields:
      type: array
      items:
        type: string
        example: deaths
      example: [
        "cases", "total_reinfections", "deaths", "deaths_cum", "recovered", "beds_occupancy", "icu_occupancy",
        "intubated", "intubated_vac", "intubated_unvac", "hospital_admissions", "hospital_discharges",
        "estimated_new_rtpcr_tests", "estimated_new_rapid_tests", "estimated_new_total_tests"
      ]
    regionalUnit:
      description: greek prefecture
      type: object
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          description: a slugged version of the normalized name of the regional unit
          example: "artas"
        department:
          type: string
          example: "Ήπειρος"
        prefecture:
          type: string
          example: "Περιφέρεια Ηπείρου"
        regional_unit_normalized:
          type: string
          description: the normalized version of the prefecture name in greek capital letters
          example: "ΑΡΤΑΣ"
        regional_unit:
          type: string
          example: "Π.Ε. Άρτας"
        pop_11:
          type: integer
          example: 38473
    regionalUnits:
      type: array
      items:
        $ref: '#/components/schemas/regionalUnit'
    municipalities:
      type: array
      items:
        $ref: '#/components/schemas/municipality'
    municipality:
      description: greek municipality
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Πάργας"
        slug:
          type: string
          description: a slugged version of municipality name
          example: "pargas"
    yearlyDeaths:
      description: deaths for a specific year and municipality
      type: object
      properties:
        year:
          type: integer
          example: 2021
        deaths:
          type: integer
          example: 100
        municipality_id:
          type: integer
          description: id of a specific municipality (see GET /municipalities for details)
          example: 12
    deathsPerMunicipality:
      type: array
      items:
        $ref: '#/components/schemas/yearlyDeaths'
    singleFieldCovidInfo:
      description: info for a specific timeline field
      type: object
      properties:
        date:
          type: string
          example: "2021-01-01"
        field:
          type: string
          example: deaths
    covidInfo:
      description: daily covid19 full info
      type: object
      properties:
        date:
          type: string
          example: "2021-01-01"
        cases:
          type: integer
          example: 50
        total_reinfections:
          type: integer
          example: 50
        deaths:
          type: integer
          example: 50
        deaths_cum:
          type: integer
          example: 50
        recovered:
          type: integer
          example: 50
        hospital_admissions:
          type: integer
          example: 50
        hospital_discharges:
          type: integer
          example: 50
        intubated:
          type: integer
          example: 50
        intubated_vac:
          type: integer
          example: 50
        intubated_unvac:
          type: integer
          example: 50
        icu_occupancy:
          type: integer
          example: 50
        beds_occupancy:
          type: integer
          example: 50
        estimated_new_rtpcr_tests:
          type: integer
          example: 50
        estimated_new_rapid_tests:
          type: integer
          example: 50
        estimated_new_total_tests:
          type: integer
          example: 50
    demographicInfo:
      description: daily covid19 demographic info per age category
      type: object
      properties:
        date:
          type: string
          example: "2021-01-01"
        category:
          type: string
          example: 18-39
        cases:
          type: integer
          example: 50
        deaths:
          type: integer
          example: 50
        intensive:
          type: integer
          example: 50
        discharged:
          type: integer
          example: 50
        hospitalized:
          type: integer
          example: 50
        hospitalized_in_icu:
          type: integer
          example: 50
        passed_away:
          type: integer
          example: 50
        recovered:
          type: integer
          example: 50
        treated_at_home:
          type: integer
          example: 50
  parameters:
    page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        example: 1
    per_page:
      in: query
      name: per_page
      required: false
      schema:
        type: integer
        example: 100